mod: reorganise imports
