+++ src/core/client/mod.rs
@@ -1,1 +1,3 @@
 pub mod client;
+
+pub use client::LabJackClient;
