+++ src/core/connection/mod.rs
@@ -1,2 +1,6 @@
 pub mod discover;
 pub mod connect;
+pub mod methods;
+
+pub use discover::*;
+pub use connect::*;
