+++ src/core/conversion/mod.rs
@@ -1,3 +1,7 @@
 pub mod adc;
 pub mod dac;
 pub mod daq;
+
+pub use adc::*;
+pub use dac::*;
+pub use daq::*;
