+++ src/core/dist.rs
@@ -1,8 +1,6 @@
 use log::{debug, warn};
-use crate::prelude::client::client::LabJackClient;
-use crate::prelude::connect::Connect;
-use crate::prelude::modbus::TcpTransport;
-use super::{discover::Discover, modbus::Error, DeviceType, LabJackDevice, LabJackSerialNumber, EMULATED_DEVICE_SERIAL_NUMBER};
+
+use crate::prelude::*;
 
 pub struct LabJack;
